+++ src/components/spartan/SDefinitionTerm/index.ts
@@ -1,1 +1,3 @@
 export { default as SDefinitionTerm } from './SDefinitionTerm.vue';
+
+export type { TDefinitionTermProps } from './types';