+++ src/components/spartan/SDropdown/index.ts
@@ -1,2 +1,4 @@
 export { default as SDropdown } from './SDropdown.vue';
 export { default as SDropdownItem } from './SDropdownItem.vue';
+
+export type { TDropdownProps, TDropdownItemProps } from './types';