+++ src/components/spartan/SLabel/index.ts
@@ -1,1 +1,3 @@
 export { default as SLabel } from './SLabel.vue';
+
+export type { TLabelProps } from './types';