+++ src/components/spartan/SLink/index.ts
@@ -1,1 +1,3 @@
 export { default as SLink } from './SLink.vue';
+
+export type { TLinkProps } from './types';