+++ src/components/spartan/SPopover/types.ts
@@ -1,5 +1,9 @@
 import type { Placement } from '@floating-ui/vue';
 
+export type TPopoverEmits = {
+    (event: 'close'): void;
+};
+
 export type TPopoverProps = {
     static?: boolean;
     offset?: number;