+++ src/components/spartan/SRadioGroup/index.ts
@@ -1,2 +1,4 @@
 export { default as SRadioGroup } from './SRadioGroup.vue';
 export { default as SRadioGroupItem } from './SRadioGroupItem.vue';
+
+export type { TRadioGroupEmits, TRadioGroupProps, TRadioGroupItemProps } from './types';