+++ src/components/spartan/SSectionTitle/index.ts
@@ -1,1 +1,3 @@
 export { default as SSectionTitle } from './SSectionTitle.vue';
+
+export type { TSectionTitleProps } from './types';