+++ src/components/spartan/SSelect/types.ts
@@ -1,5 +1,9 @@
 import { type TRounded } from '@/helpers';
 
+export type TSelectEmits = {
+    (e: 'update:modelValue', value: string | number | undefined): void;
+};
+
 export type TSelectProps = {
     disabled: boolean;
     error: boolean;