+++ src/components/spartan/SSteps/types.ts
@@ -1,0 +1,8 @@
+export type TStepsProps = {
+    steps: Array<{
+        name: string;
+        href: string;
+        description: string;
+        status: 'complete' | 'current' | 'upcoming';
+    }>;
+};