+++ src/components/spartan/STab/index.ts
@@ -1,2 +1,4 @@
 export { default as STab } from './STab.vue';
 export { default as STabItem } from './STabItem.vue';
+
+export type { TTabProps, TTabEmits, TTabItemProps } from './types';