+++ src/components/spartan/STab/types.ts
@@ -1,14 +1,15 @@
-import type { Component, FunctionalComponent } from "vue";
+import type { Component, FunctionalComponent } from 'vue';
 
 export type TTab = {
     path: string;
-    setActive: (value: boolean) => void
+    setActive: (value: boolean) => void;
 };
 
 export type TTabProps = {
     variant: 'underline' | 'pills';
     modelValue: string;
 };
+
 export type TTabEmits = {
     (event: 'update:modelValue', value?: string): void;
 };